webapp: removing duplicate code - moving it a to single utils file
